+++ src/engine/engine.hpp
@@ -6,10 +6,10 @@ class IScene;
 
 namespace engine {
 	struct Timestep {
-		uint64_t time;
-		uint64_t dt;
+		float time;
+		float dt;
 		uint64_t currenttime;
-		uint64_t accumulator;
+		float accumulator;
 
 		uint32_t remaining_ticks;
 	};
